requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.4.0
python-dotenv==1.0.0
numpy>=1.24.0,<2.0.0
rembg==2.0.67
transformers>=4.36.0
torch>=2.6.0
torchvision>=0.19.0
kornia
timm
--extra-index-url https://download.pytorch.org/whl/cpu